+++ frontend/src/components/multi-step-forms/StepSlider.vue
