+++ frontend/src/components/multi-step-forms/application/steps/ApplicationStep.vue
@@ -1,0 +1,13 @@
+<template>
+    <h1>Hello It's me, Mario!</h1>
+</template>
+
+<script>
+export default {
+    name: 'ApplicationStep'
+}
+</script>
+
+<style scoped lang="scss">
+
+</style>
